requirements.txt
absl-py==1.3.0
alembic==1.8.1
altair==4.2.0
anaconda==0.0.1.1
app==0.0.1
appdirs==1.4.4
appnope==0.1.2
arabic-reshaper==2.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asn1crypto==1.5.1
astor==0.8.1
astroid==2.15.4
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
blinker==1.4
blis==0.7.9
bs4==0.0.1
cachetools==5.0.0
catalogue==2.0.8
category-encoders==2.6.0
certifi==2019.3.9
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.1.3
cmake==3.25.0
colorlover==0.3.0
confection==0.0.4
cryptography==38.0.3
cssselect2==0.7.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.1
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-mathfilters==1.0.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dnspython==2.2.1
email-validator==1.1.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.0.12
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-Scss==0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==23.3.3
fonttools==4.28.5
future==0.18.2
gast==0.4.0
gensim==4.3.1
geocoder==1.38.1
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.15.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.51.1
h5py==3.8.0
html5lib==1.1
idna==3.3
inflect==6.0.2
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.12.0
itsdangerous==2.1.2
jax==0.4.8
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.12.0
kiwisolver==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.9.0
libclang==16.0.0
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
ml-dtypes==0.0.4
mpython==1.0
murmurhash==1.0.9
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
node==1.2.1
notebook==6.4.7
npm==0.1.1
ntplib==0.4.0
numpy==1.21.6
oauthlib==3.2.2
odict==1.9.0
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
optional-django==0.1.0
oscrypto==1.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathy==0.10.1
patsy==0.5.3
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2023.2.4
platformdirs==2.4.1
plotly==5.14.0
plumber==1.7
preshed==3.0.8
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.7.1
Pygments==2.11.2
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyJWT==2.6.0
pylint==2.17.3
Pympler==1.0.1
pyparsing==3.0.6
PyPDF3==1.0.6
pyrsistent==0.18.0
pyScss==1.4.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2018.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
qrcode==7.3.1
ratelim==0.1.6
razorpay==1.3.0
react==4.3.0
regex==2022.9.13
reportlab==3.6.12
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
soupsieve==2.3.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.44
sqlparse==0.4.4
srsly==2.4.6
stack-data==0.1.3
statsmodels==0.13.5
streamlit==1.4.0
stripe==5.4.0
svglib==1.4.1
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
termcolor==2.2.0
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
thinc==8.1.9
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==1.2.3
tomlkit==0.11.7
toolz==0.11.2
torch==1.13.0
torchvision==0.14.0
tornado==6.1
tqdm==4.64.1
traitlets==5.1.1
typer==0.7.0
typing_extensions==4.5.0
tzdata==2021.5
tzlocal==4.1
uritools==4.0.0
urllib3==1.24.2
vaderSentiment==3.3.2
validators==0.18.2
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wasabi==1.1.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.5.2
wordcloud==1.8.2.2
wrapt==1.12.1
WTForms==3.0.1
xgboost==1.6.2
xhtml2pdf==0.2.8
zope.component==6.0
zope.deferredimport==4.4
zope.deprecation==5.0
zope.event==4.6
zope.hookable==5.4
zope.interface==6.0
zope.lifecycleevent==4.4
zope.proxy==5.0.0
